+++ src/extensions/techno/technoext_hooks.cpp
@@ -2764,20 +2764,22 @@ static bool Is_Valid_Target_Zone(const TechnoClass* techno, AbstractClass * targ
 
 
 /**
- * #discussions-1450 
- * Patch that allows modders and mappers to customize repair cap and repair rate across technos,
- * rather than relying on values that have other side effects.
- * Technos can have individual repair cap and rates specified for them.
- * If they don't have those keys, then those values fall back to the game-wide customizable keys.
- * If those also aren't specified either, then the game falls back to the keys it used in vanilla.
+ *  #discussions-1450 
+ *  Patch that allows customizing repair cap and repair rate across technos,
+ *  rather than relying on values that have other side effects.
+ *  Technos can have individual repair cap and rates specified for them.
+ *  If they don't have those keys, then those values fall back to the game-wide customizable keys.
+ *  If those also aren't specified either, then the game falls back to the keys it used in vanilla.
  *
- * @author: JoyfulShush
+ *  @author: JoyfulShush
  */
 bool TechnoClassExt::_Should_Self_Heal_Now() const
 {    
     auto technotypeext = Extension::Fetch(TClass);
 
-    // Prevents overhealing - allowing Strength to get above 100%.
+    /**
+     *  Prevents overhealing - allowing Strength to get above 100%.
+     */
     if (Get_Health_Ratio() >= 1) {
         return false;
     }
@@ -2786,15 +2788,15 @@ bool TechnoClassExt::_Should_Self_Heal_Now() const
         if (!Has_Ability(ABILITY_SELF_HEAL)) {
             return false;
         }
-    }    
+    }
 
     /**
-     * Determine the repair rate of this techno. 
-     * If the techno has its own repair rate, it is used. 
-     * Otherwise, the game-wide repair rate is used.
-     * If both are not specified, then the game falls back to the original RepairRate key used by this function in vanilla.     
-     */    
-    float repair_rate;
+     *  Determine the repair rate of this techno. 
+     *  If the techno has its own repair rate, it is used. 
+     *  Otherwise, the game-wide repair rate is used.
+     *  If both are not specified, then the game falls back to the original RepairRate key used by this function in vanilla.     
+     */
+    double repair_rate;
 
     if (technotypeext->SelfHealingRate != -1) {
         repair_rate = technotypeext->SelfHealingRate;
@@ -2811,10 +2813,10 @@ bool TechnoClassExt::_Should_Self_Heal_Now() const
     }
 
     /**
-     * Determines the repair cap (in percentage) of this techno. A techno cannot repair itself beyond its repair cap.
-     * If the techno has its own repair cap, it is used.
-     * Otherwise, the game-wide repair cap is used.
-     * If both are not specified, then the game falls back to the original ConditionYellow key used by this function in vanilla.
+     *  Determines the repair cap (in percentage) of this techno. A techno cannot repair itself beyond its repair cap.
+     *  If the techno has its own repair cap, it is used.
+     *  Otherwise, the game-wide repair cap is used.
+     *  If both are not specified, then the game falls back to the original ConditionYellow key used by this function in vanilla.
      */
     float repair_cap;
     
@@ -3179,6 +3181,7 @@ bool TechnoClassExt::_Evaluate_Object(ThreatType method, int mask, int range, Te
     return false;
 }
 
+
 /**
  *  Fixes a bug where placed buildings are not revealed for allies.
  *
@@ -3196,8 +3199,6 @@ DEFINE_HOOK(0x0062AB5E, _TechnoClass_Revealed_Look_For_Allies_Patch, 0)
     // trigger TEVENT_DISCOVERED
     return 0x0062AB68;
 }
-
-
 
 
 /**
